Restructured player in a bid to speed it up
(which failed).  Now, instead of always initialising the player with Either a recording or not, we only use the player if we actually have a recording.  i.e. the calling program is responsible for displaying the widget if it has a recording available and displaying nothing if not.

Unfortunately, for some reason, when the player is implemented as a Pux module (using the standard TEA architecture) its responsiveness slows down and we no longer are able to play the tune at anything like its intended tempo.  I presume that this is a Pux bug related to unneccesary re-rendering of the widget.