Major refactoring - reduce dependencies
The base behaviour of the player has been hived off into a separate module - purescript-soundfont-player from which this player inherits.  This one is now restricted to just playing MIDI (ABC support will be removed to a separate player).

This means that there is no longer a single player with multiple unnecessary dependencies, but dedicated players for each environment with properly scoped dependencies.